# src/main/java/it/bitify/libreria/model/dto/IdStudentBookDTO.java
package it.bitify.libreria.model.dto;

public class IdStudentBookDTO {
    private Long idStudent;
    private Long idBook;

    public Long getIdStudent() {
        return idStudent;
    }

    public void setIdStudent(Long idStudent) {
        this.idStudent = idStudent;
    }

    public Long getIdBook() {
        return idBook;
    }

    public void setIdBook(Long idBook) {
        this.idBook = idBook;
    }
}