# src/main/resources/static/openapi.yaml
openapi: 3.1.0
info:
  title: Libreria API definition
  version: v1
servers:
- url: http://localhost:8080
  description: base server url
...
paths:
...
  /api/v1/student:
    get:
      tags:
      - get-all
      - Student
      summary: Retrieve all students
      description: Retrieve all student objects. The response is a list of student
        objects
      operationId: getAll
      parameters:
        - $ref: '#/components/parameters/PageParam'
        - $ref: '#/components/parameters/SizeParam'
        - $ref: '#/components/parameters/AscParam'
        - $ref: '#/components/parameters/SortByParam'
      responses:
        "200":
          description: Retrieved all student
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PageStudent"
    put:
      tags:
      - Student
      - put
      summary: Modify student
      description: Modify a student object by specifying its id
      operationId: editStudent
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/Student"
        required: true
      responses:
        "200":
          description: Student modified
        "400":
          description: Id type mismatch
        "404":
          description: Student not found
        "500":
          description: Data integrity violation
    post:
      tags:
      - post
      - Student
      summary: Add student
      description: Add a student object to the database
      operationId: addstudent
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/Student"
        required: true
      responses:
        "200":
          description: Student added to the database
        "500":
          description: Data integrity violation
  /api/v1/student/{id}:
    get:
      tags:
        - get
        - Student
      summary: Retrieve student by id
      description: Retrieve a student object by specifying its id. The response is
        a student object
      operationId: getstudentById
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        "200":
          description: Student found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Student"
        "400":
          description: Id type mismatch
        "404":
          description: Student not found
    delete:
      tags:
        - Student
        - delete
      summary: Delete student by id
      description: Delete a student object by specifying its id
      operationId: deleteStudent
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        "200":
          description: Student deleted
        "400":
          description: Id type mismatch
        "404":
          description: Student not found
        "500":
          description: Data integrity violation
  /api/v1/student/without-loan:
    get:
      tags:
        - get-all
        - Student
      summary: Retrieve students without loans
      description: Retrieve a list of student objects without loans
      operationId: getStudentWithoutLoan
      parameters:
        - $ref: '#/components/parameters/PageParam'
        - $ref: '#/components/parameters/SizeParam'
        - $ref: '#/components/parameters/AscParam'
        - $ref: '#/components/parameters/SortByParam'
      responses:
        "200":
          description: Student found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PageStudent"
  /api/v1/student/by-class:
    get:
      tags:
        - get-all
        - Student
      summary: Retrieve students by class
      description: Retrieve a list of student objects by specifying a class
      operationId: getStudentByClass
      parameters:
        - name: schoolClass
          in: query
          required: true
          schema:
            type: string
        - $ref: '#/components/parameters/PageParam'
        - $ref: '#/components/parameters/SizeParam'
        - $ref: '#/components/parameters/AscParam'
        - $ref: '#/components/parameters/SortByParam'
      responses:
        "200":
          description: Student found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PageStudent"
  /api/v1/student/by-card-release:
    get:
      tags:
        - get-all
        - Student
      summary: Retrieve students by card release date
      description: Retrieve a list of student objects by specifying a card release date
      operationId: getStudentByCardRelease
      parameters:
        - name: start
          in: query
          required: false
          schema:
            type: string
            format: date
        - name: end
          in: query
          required: false
          schema:
            type: string
            format: date
        - $ref: '#/components/parameters/PageParam'
        - $ref: '#/components/parameters/SizeParam'
        - $ref: '#/components/parameters/AscParam'
        - $ref: '#/components/parameters/SortByParam'
      responses:
        "200":
          description: Student found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PageStudent"
  /api/v1/student/min-loan:
    get:
      tags:
        - get-all
        - Student
      summary: Retrieve students with more loans then specified
      description: Retrieve a list of student objects by specifying a minimum number of loans
      operationId: getStudentByMinLoan
      parameters:
        - name: minLoan
          in: query
          required: true
          schema:
            type: integer
            format: int32
        - $ref: '#/components/parameters/PageParam'
        - $ref: '#/components/parameters/SizeParam'
        - $ref: '#/components/parameters/AscParam'
        - $ref: '#/components/parameters/SortByParam'
      responses:
        "200":
          description: Student found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PageNameSurnameLoansDTO"
  /api/v1/student/by-category-name:
    get:
      tags:
        - get-all
        - Student
      summary: Retrieve students by specifying a category
      description: Retrieve a list of student objects by specifying a category
      operationId: getStudentByCategoryName
      parameters:
        - name: categoryName
          in: query
          required: true
          schema:
            type: string
        - $ref: '#/components/parameters/PageParam'
        - $ref: '#/components/parameters/SizeParam'
        - $ref: '#/components/parameters/AscParam'
        - $ref: '#/components/parameters/SortByParam'
      responses:
        "200":
          description: Student found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PageStudent"
  /api/v1/student/by-loan-count:
    get:
      tags:
        - get
        - Student
      summary: Retrieve students ordered by number of loans
      description: Retrieve a list of student objects ordered by number of loans
      operationId: getStudentByLoanCount
      parameters:
        - $ref: '#/components/parameters/PageParam'
        - $ref: '#/components/parameters/SizeParam'
        - $ref: '#/components/parameters/AscParam'
        - $ref: '#/components/parameters/SortByParam'
      responses:
        "200":
          description: Student found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PageStudent"
  /api/v1/student/loan-book:
    post:
      tags:
        - Student
        - Post
      summary: Loan book
      description: Loan a book if it isn't already lent
      operationId: loanBook
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/IdStudentBookDTO"
      responses:
        "200":
          description: Book lent
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Loan"
        "400":
          description: Book already lent
        "404":
          description: Entity not found
  /api/v1/student/return-book:
    post:
      tags:
        - Student
        - Post
      summary: Return book
      description: Return a book if it is already lent
      operationId: returnBook
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/IdStudentBookDTO"
      responses:
        "200":
          description: Book returned
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Loan"
        "404":
          description: Entity not found
...
components:
  schemas:
    Card:
      type: object
      properties:
        id:
          type: integer
          format: int64
        releaseDate:
          type: string
          format: date
    Student:
      type: object
      properties:
        id:
          type: integer
          format: int64
        name:
          type: string
        surname:
          type: string
        class:
          type: string
        email:
          type: string
        Card:
          $ref: "#/components/schemas/Card"
    Category:
      type: object
      properties:
        id:
          type: integer
          format: int64
        name:
          type: string
        books:
          type: array
          items:
            $ref: "#/components/schemas/Book"
    Book:
      type: object
      properties:
        id:
          type: integer
          format: int64
        year:
          type: integer
          format: int32
        category:
          $ref: "#/components/schemas/Category"
        title:
          type: string
        author:
          type: string
    Loan:
      type: object
      properties:
        id:
          type: integer
          format: int64
        startDate:
          type: string
          format: date
        endDate:
          type: string
          format: date
        student:
          $ref: "#/components/schemas/Student"
        Book:
          $ref: "#/components/schemas/Book"
...
    IdStudentBookDTO:
      type: object
      properties:
        idStudent:
          type: integer
          example: 5
        idBook:
          type: integer
          example: 7
      required:
        - idStudent
        - idBook
...
    NameSurnameLoansDTO:
      type: object
      properties:
        name:
          type: string
          example: Luca
        surname:
          type: string
          example: rossi
        loans:
          type: integer
          example: 7
      required:
        - name
        - surname
        - loans
    PageStudent:
      type: object
      properties:
        content:
          type: array
          items:
            $ref: '#/components/schemas/Student'
        totalElements:
          type: integer
        totalPages:
          type: integer
        size:
          type: integer
        number:
          type: integer
...
    PageNameSurnameLoansDTO:
      type: object
      properties:
        content:
          type: array
          items:
            $ref: '#/components/schemas/NameSurnameLoansDTO'
        totalElements:
          type: integer
        totalPages:
          type: integer
        size:
          type: integer
        number:
          type: integer
  parameters:
    PageParam:
      name: page
      in: query
      required: false
      schema:
        type: integer
        default: 0
    SizeParam:
      name: size
      in: query
      required: false
      schema:
        type: integer
        default: 5
    AscParam:
      name: ascending
      in: query
      required: false
      schema:
        type: boolean
        default: true
    SortByParam:
      name: sortBy
      in: query
      required: false
      schema:
        type: string
        default: id